+++ src/model/mod.rs
@@ -62,11 +62,10 @@ let data = r#"
 ```
 
 */
-
-pub mod decimal;
 pub mod manifest;
 pub mod partition;
 pub mod schema;
 pub mod snapshot;
 pub mod sort;
 pub mod table;
+pub mod types;
