+++ src/model/types/struct_field.rs
@@ -1,0 +1,8 @@
+/*!
+ * Struct type in iceberg
+ */
+
+use super::Nullable;
+
+/// Struct field is tuple of name, value, and optional documentation.
+pub type StructField = (String, Nullable, Option<String>);
